+++ host/vulkan/VkDecoderGlobalState.cpp
@@ -489,10 +489,10 @@ class VkDecoderGlobalState::Impl {
                 // regardless, we might hit a Vulkan validation error because the new image might
                 // have the "usage" flag that is unsuitable to bind to descriptors.
                 std::vector<std::pair<int, int>> validWriteIndices;
-                for (int bindingIdx = 0; bindingIdx < descriptorSetInfo.allWrites.size();
+                for (int bindingIdx = 0; bindingIdx < (int)descriptorSetInfo.allWrites.size();
                      bindingIdx++) {
                     for (int bindingElemIdx = 0;
-                         bindingElemIdx < descriptorSetInfo.allWrites[bindingIdx].size();
+                         bindingElemIdx < (int)descriptorSetInfo.allWrites[bindingIdx].size();
                          bindingElemIdx++) {
                         const auto& entry = descriptorSetInfo.allWrites[bindingIdx][bindingElemIdx];
                         if (entry.writeType == DescriptorSetInfo::DescriptorWriteType::Empty) {
@@ -500,7 +500,7 @@ class VkDecoderGlobalState::Impl {
                         }
                         int dependencyObjCount =
                             descriptorDependencyObjectCount(entry.descriptorType);
-                        if (entry.alives.size() < dependencyObjCount) {
+                        if ((int)entry.alives.size() < dependencyObjCount) {
                             continue;
                         }
                         bool isValid = true;
@@ -752,7 +752,7 @@ class VkDecoderGlobalState::Impl {
                     VkDescriptorSetLayout boxedLayout = (VkDescriptorSetLayout)stream->getBe64();
                     layouts.push_back(unbox_VkDescriptorSetLayout(boxedLayout));
                     uint64_t validWriteCount = stream->getBe64();
-                    for (int write = 0; write < validWriteCount; write++) {
+                    for (uint64_t write = 0; write < validWriteCount; write++) {
                         uint32_t binding = stream->getBe32();
                         uint32_t arrayElement = stream->getBe32();
                         DescriptorSetInfo::DescriptorWriteType writeType =
@@ -3637,7 +3637,7 @@ class VkDecoderGlobalState::Impl {
         for (size_t i = 0; i < setInfo.bindings.size(); i++) {
             VkDescriptorSetLayoutBinding dslBinding = setInfo.bindings[i];
             int bindingIdx = dslBinding.binding;
-            if (setInfo.allWrites.size() <= bindingIdx) {
+            if ((int)setInfo.allWrites.size() <= bindingIdx) {
                 setInfo.allWrites.resize(bindingIdx + 1);
             }
             setInfo.allWrites[bindingIdx].resize(dslBinding.descriptorCount);
@@ -6238,7 +6238,7 @@ class VkDecoderGlobalState::Impl {
         if (!m_emu->features.GuestVulkanOnly.enabled) {
             {
                 std::lock_guard<std::mutex> lock(mMutex);
-                for (int i = 0; i < submitCount; i++) {
+                for (uint32_t i = 0; i < submitCount; i++) {
                     for (int j = 0; j < getCommandBufferCount(pSubmits[i]); j++) {
                         VkCommandBuffer cmdBuffer = getCommandBuffer(pSubmits[i], j);
                         CommandBufferInfo* cmdBufferInfo =
@@ -6322,7 +6322,7 @@ class VkDecoderGlobalState::Impl {
         {
             std::lock_guard<std::mutex> lock(mMutex);
             std::unordered_set<HandleType> imageBarrierColorBuffers;
-            for (int i = 0; i < submitCount; i++) {
+            for (uint32_t i = 0; i < submitCount; i++) {
                 for (int j = 0; j < getCommandBufferCount(pSubmits[i]); j++) {
                     VkCommandBuffer cmdBuffer = getCommandBuffer(pSubmits[i], j);
                     CommandBufferInfo* cmdBufferInfo =
@@ -6350,7 +6350,7 @@ class VkDecoderGlobalState::Impl {
         {
             std::lock_guard<std::mutex> lock(mMutex);
             // Update image layouts
-            for (int i = 0; i < submitCount; i++) {
+            for (uint32_t i = 0; i < submitCount; i++) {
                 for (int j = 0; j < getCommandBufferCount(pSubmits[i]); j++) {
                     VkCommandBuffer cmdBuffer = getCommandBuffer(pSubmits[i], j);
                     CommandBufferInfo* cmdBufferInfo =
@@ -6370,15 +6370,15 @@ class VkDecoderGlobalState::Impl {
             // Update latestUse for all wait/signal semaphores, to ensure that they
             // are never asynchronously destroyed before the queue submissions referencing
             // them have completed
-            for (int i = 0; i < submitCount; i++) {
-                for (int j = 0; j < getWaitSemaphoreCount(pSubmits[i]); j++) {
+            for (uint32_t i = 0; i < submitCount; i++) {
+                for (uint32_t j = 0; j < getWaitSemaphoreCount(pSubmits[i]); j++) {
                     SemaphoreInfo* semaphoreInfo =
                         android::base::find(mSemaphoreInfo, getWaitSemaphore(pSubmits[i], j));
                     if (semaphoreInfo) {
                         semaphoreInfo->latestUse = queueCompletedWaitable;
                     }
                 }
-                for (int j = 0; j < getSignalSemaphoreCount(pSubmits[i]); j++) {
+                for (uint32_t j = 0; j < getSignalSemaphoreCount(pSubmits[i]); j++) {
                     SemaphoreInfo* semaphoreInfo =
                         android::base::find(mSemaphoreInfo, getSignalSemaphore(pSubmits[i], j));
                     if (semaphoreInfo) {
@@ -7112,7 +7112,7 @@ class VkDecoderGlobalState::Impl {
             // Track the Colorbuffers that would be written to.
             // It might be better to check for VK_QUEUE_FAMILY_EXTERNAL in pipeline barrier.
             // But the guest does not always add it to pipeline barrier.
-            for (int i = 0; i < pCreateInfo->attachmentCount; i++) {
+            for (uint32_t i = 0; i < pCreateInfo->attachmentCount; i++) {
                 auto* imageViewInfo = android::base::find(mImageViewInfo, pCreateInfo->pAttachments[i]);
                 if (imageViewInfo->boundColorBuffer.has_value()) {
                     framebufferInfo.attachedColorBuffers.push_back(
